fix(categories): :bug: Use hashes to avoid installation issues.
I was trying to have a selector for cpu vs gpu - to include pytorch+cpu when I select cpu and to include the heavy gpu version only when I select it. Using the definition in the image.  It always seemed to pick up the pytorch Cpu libs and it was really confusing me as it would include the nvidia packages as well. This was until I ensured I exported the hashes into the requirements file.  Without this the installation was getting very confused.  But with the hashes it does the right thing.